backend/gpt_researcher_server/requirements.txt
fastapi
uvicorn
gpt-researcher>=0.9.0
python-dotenv
aiohttp
gunicorn
langchain-openai
langchain-community
# Development tools for type checking and formatting
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0